+++ forui/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 0.11.1 (Next)
+## 0.11.1
 
+* Add optional named parameters with their default values to CLI generated styles. 
 * Fix `FTileStyle.pressable` not changing background color on press & hold.
-
+* Fix typo in CLI generated styles' documentation.
 
 
 ## 0.11.0